# PERF: Fix performance regression due to CoW ref tracking