@phofl 2023-02-14
2 things are not behind CoW:
- creating this object
- adding the ref

With this implementation, the more costly part is checking the ref, which is hidden behind CoW. There is no real efficient way of hiding everything (at least I can't see one right now)