@phofl 2023-02-14
This was cheap with the WeakSet when I checked where we spent the most time