@phofl 2023-02-15
Yes this is how it was implemented initially

You are correct, that it slows down the has_reference check. Two points why I think this is a better idea right now:

- it only hits in case of CoW is enabled, e.g. the default is not touched by this
- the list is small most of the time, what you said basically