@wangwillian0 2023-09-05
Hi @phofl, I think this commit may be related to the problem at https://github.com/pandas-dev/pandas/issues/54352. Each weakref.ref ocupies 72 bytes of memory even after it's a dead , which accumulates invisibly on tons of operations (e.g. Index.view called during df.copy). The gap is probably a little closer too if you consider the extreme case of the weakref.WeakSet() having always length 1 when the current list version keeps allocating memory. Do you think it's worth it to revert this commit?