@phofl 2023-09-05
We can clear the list more often, that would be an approach worth considering as well